feat: introduce omit flag
