+++ test/fixtures/array.prototype.foreach.js
@@ -1,0 +1,2 @@
+[].map(() => {});
+[].forEach(() => {});
